Move utils to migrations_utils.go
